+++ lib/common/constants.dart
@@ -56,10 +56,6 @@ class Constants {
   /// The offset in minutes for "consume-by" field.
   static const foodConsumeByMinutesOffset = 30;
 
-  // ZOB-234 Use shorter confirmation time for personal delivery
-  static const pickupConfirmationTimeDefault = 45;
-  static const pickupConfirmationTimePersonal = 20;
-
   // ZOB-305 Food temperature related constants
   static const foodTemperatureMin = 50;
   static const foodTemperatureMax = 100;